refactor(masonry): [brick] rename newPath to path2
